SWATCH-2161: Removed unused InventoryAccountUsageCollector.collect
This method was mainly used for testing purposes, so part of its functionality has been moved on InventoryAccountUsageCollectorTallyTest.
